requirements.txt
aiofiles==22.1.0
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
httptools==0.5.0
multidict==6.0.2
pendulum==2.1.2
python-dateutil==2.8.2
pytz==2022.4
pytzdata==2020.1
sanic==22.9.0
sanic-routing==22.8.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
websockets==10.3
